Add TypeScript version of extensions_putting_together.py (#6597)
* Add TypeScript version of extensions_putting_together.py

* update user guide with typescript example
